# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OmniReferral API",
        "description": "REST API for the OmniReferral real estate referral and lead management platform.\n\nProvides authentication, user management, realtor profiles, GoHighLevel CRM integration,\nonboarding webhooks, and administrative functions.\n\n## Authentication\nMost endpoints require a Bearer token obtained via the Login or Register endpoints.\nPass the token in the `Authorization` header as `Bearer <token>`.\n\n## Webhooks\nGoHighLevel webhook endpoints are public (CSRF exempt) and authenticated via\nthe `X-OmniReferral-Webhook` header secret.",
        "contact": {
            "email": "admin@omnireferrals.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://omnireferrals.com",
            "description": "OmniReferral Production API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/login": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Show login form",
                "operationId": "749bdaf6374ca2f6fdb149d27225370e",
                "responses": {
                    "200": {
                        "description": "Login page"
                    }
                }
            },
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user and start session",
                "description": "Logs in a user with workspace role, email, and password. Starts a web session.",
                "operationId": "ec9c379d2a959de66404708dd4a9322d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect to dashboard on success"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/forgot-password": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Show forgot password form",
                "operationId": "b52f4f77cb19a399fc4037fffb816b9d",
                "responses": {
                    "200": {
                        "description": "Forgot password page"
                    }
                }
            },
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send password reset link",
                "description": "Sends a password reset email to the given address.",
                "operationId": "60e8a8c89f9c4d7c4e50c9ee676b1ada",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect with success or error message"
                    }
                }
            }
        },
        "/reset-password/{token}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Show password reset form",
                "operationId": "66bf18ecbd1c28794d502de617a76578",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reset password page"
                    }
                }
            }
        },
        "/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password with token",
                "description": "Consumes a password reset token and sets a new password.",
                "operationId": "3c793085ef8a11bc3209c9bc94792715",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect to login on success"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log out and invalidate session",
                "operationId": "cc14526f9abd470d63047b1c05aef0c3",
                "responses": {
                    "302": {
                        "description": "Redirect to home"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "required": [
                    "role",
                    "email",
                    "password"
                ],
                "properties": {
                    "role": {
                        "description": "Workspace role",
                        "type": "string",
                        "enum": [
                            "agent",
                            "admin",
                            "staff"
                        ]
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password"
                    },
                    "remember": {
                        "type": "boolean",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ForgotPasswordRequest": {
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            },
            "ResetPasswordRequest": {
                "required": [
                    "token",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "token": {
                        "description": "Password reset token from email",
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 8
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "sanctum",
                "description": "Enter Sanctum token in format: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}